requirements.txt
breadability==0.1.20
chardet==2.3.0
contextlib2==0.4.0
cssselect==0.9.1
Django==1.7.1
django-cors-headers==1.0.0
django-simpleapi==1.3.0
django-taggit==0.12.2
docopt==0.6.2
gnureadline==6.3.3
lxml==3.4.1
mock==1.0.1
pytz==2014.10
PyYAML==3.11
requests==2.5.1
responses==0.3.0
schematics==1.0.post0
six==1.9.0
vcrpy==1.1.4
w3lib==1.10.0
wrapt==1.10.2
dj-database-url==0.3.0
gunicorn==19.1.1
django-postgrespool==0.3.0
SQLAlchemy==0.9.8
psycopg2==2.5.4
whitenoise==1.0.6
